requirements.txt
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
aiosqlite==0.19.0
attrs==23.1.0
beautifulsoup4==4.12.2
et-xmlfile==1.1.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
multidict==6.0.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
prompt==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
yarl==1.9.3
